+++ packages/nimble-components/src/icons/looping.ts
@@ -1,0 +1,19 @@
+import { statusLooping16X16 } from '@ni/nimble-tokens/dist-icons-esm/nimble-icons-inline';
+import { Icon, registerIcon } from '../icon-base';
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'nimble-looping-icon': LoopingIcon;
+    }
+}
+
+/**
+ *  The icon component for the 'looping' icon
+ */
+export class LoopingIcon extends Icon {
+    public constructor() {
+        super(statusLooping16X16);
+    }
+}
+
+registerIcon('looping-icon', LoopingIcon);