+++ packages/nimble-components/src/theme-provider/index.ts
@@ -1,46 +1,84 @@
 import {
     DesignSystem,
-    DesignTokenValue,
+    DesignToken,
     FoundationElement
 } from '@microsoft/fast-foundation';
 import { attr } from '@microsoft/fast-element';
+import { Direction } from '@microsoft/fast-web-utilities';
 import { template } from './template';
 import { styles } from './styles';
-import { theme } from './design-tokens';
-import type { NimbleTheme } from './types';
-
-export type { NimbleThemeProvider };
+import { Theme } from './types';
 
 declare global {
     interface HTMLElementTagNameMap {
-        'nimble-theme-provider': NimbleThemeProvider;
+        'nimble-theme-provider': ThemeProvider;
     }
 }
 
+// Not represented as a CSS Custom Property, instead available
+// as an attribute of theme provider.
+export const direction = DesignToken.create<Direction>({
+    name: 'direction',
+    cssCustomPropertyName: null
+}).withDefault(Direction.ltr);
+
+export const theme = DesignToken.create<Theme>({
+    name: 'theme',
+    cssCustomPropertyName: null
+}).withDefault(Theme.Light);
+
 /**
- * The NimbleThemeProvider implementation. Add this component to the page and set its `theme` attribute to control
+ * The ThemeProvider implementation. Add this component to the page and set its `theme` attribute to control
  * the values of design tokens that provide colors and fonts as CSS custom properties to any descendant components.
  * @internal
  */
-class NimbleThemeProvider extends FoundationElement {
+export class ThemeProvider extends FoundationElement {
+    @attr({
+        attribute: 'direction'
+    })
+    public direction!: Direction;
+
     @attr({
         attribute: 'theme'
     })
-    public theme!: NimbleTheme;
+    public theme!: Theme;
 
-    public themeChanged(
-        _prev: NimbleTheme | undefined,
-        next: NimbleTheme | undefined
+    public directionChanged(
+        _prev: Direction | undefined,
+        next: Direction | undefined
     ): void {
         if (next !== undefined && next !== null) {
-            theme.setValueFor(this, next as DesignTokenValue<NimbleTheme>);
+            direction.setValueFor(this, next);
+        } else {
+            direction.deleteValueFor(this);
+        }
+    }
+
+    public themeChanged(
+        _prev: Theme | undefined,
+        next: Theme | undefined
+    ): void {
+        if (next !== undefined && next !== null) {
+            theme.setValueFor(this, next);
         } else {
             theme.deleteValueFor(this);
         }
     }
+
+    public connectedCallback(): void {
+        super.connectedCallback();
+
+        if (!this.direction) {
+            this.direction = Direction.ltr;
+        }
+
+        if (!this.theme) {
+            this.theme = Theme.Light;
+        }
+    }
 }
 
-const nimbleDesignSystemProvider = NimbleThemeProvider.compose({
+const nimbleDesignSystemProvider = ThemeProvider.compose({
     baseName: 'theme-provider',
     styles,
     template