+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,43 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Thu, 20 Jan 2022 21:30:14 GMT and should not be manually modified.
+This log was last generated on Fri, 04 Feb 2022 23:44:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.0.0-beta.149
+
+Fri, 04 Feb 2022 23:44:50 GMT
+
+### Changes
+
+- Name consistency
+- **Breaking change**: NimbleThemeProvider renamed to ThemeProvider. NimbleTheme renamed to Theme. Removed passwordRevealFilter design token. (rajsite@users.noreply.github.com)
+- Bump @ni/nimble-components to v1.0.0-beta.129
+
+## 1.0.0-beta.148
+
+Wed, 02 Feb 2022 16:29:26 GMT
+
+### Changes
+
+- Bump @ni/nimble-components to v1.0.0-beta.128
+
+## 1.0.0-beta.147
+
+Tue, 01 Feb 2022 17:09:06 GMT
+
+### Changes
+
+- Bump @ni/nimble-components to v1.0.0-beta.127
+
+## 1.0.0-beta.146
+
+Fri, 28 Jan 2022 21:45:26 GMT
+
+### Changes
+
+- Adds documentation and CSS/SCSS files for using theme-aware design tokens in Angular applications (rajsite@users.noreply.github.com)
+- Bump @ni/nimble-components to v1.0.0-beta.126
 
 ## 1.0.0-beta.145